+++ sdk/keyvault/azure-keyvault-keys/azure/keyvault/keys/crypto/__init__.py
@@ -4,7 +4,8 @@
 # ------------------------------------
 from .enums import EncryptionAlgorithm, KeyWrapAlgorithm, SignatureAlgorithm
 from ._models import DecryptResult, EncryptResult, SignResult, UnwrapKeyResult, VerifyResult, WrapKeyResult
-from .client import CryptographyClient
+from ._client import CryptographyClient
+from ._enums import EncryptionAlgorithm, KeyWrapAlgorithm, SignatureAlgorithm
 
 
 __all__ = [
