+++ sdk/keyvault/azure-keyvault-keys/azure/keyvault/keys/crypto/_client.py
@@ -2,13 +2,14 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 # ------------------------------------
+import six
 from azure.core.exceptions import AzureError, HttpResponseError
-from azure.core.tracing.decorator_async import distributed_trace_async
-from azure.keyvault.keys.models import KeyVaultKey
-from azure.keyvault.keys._shared import AsyncKeyVaultClientBase, parse_vault_id
+from azure.core.tracing.decorator import distributed_trace
 
-from .. import DecryptResult, EncryptResult, SignResult, VerifyResult, UnwrapKeyResult, WrapKeyResult
-from .._internal import EllipticCurveKey, RsaKey
+from . import DecryptResult, EncryptResult, SignResult, VerifyResult, UnwrapKeyResult, WrapKeyResult
+from ._internal import EllipticCurveKey, RsaKey
+from .._models import KeyVaultKey
+from .._shared import KeyVaultClientBase, parse_vault_id
 
 try:
     from typing import TYPE_CHECKING
@@ -19,21 +20,21 @@ if TYPE_CHECKING:
     # pylint:disable=unused-import
     from typing import Any, Optional, Union
     from azure.core.credentials import TokenCredential
-    from .. import EncryptionAlgorithm, KeyWrapAlgorithm, SignatureAlgorithm
-    from .._internal import Key as _Key
+    from . import EncryptionAlgorithm, KeyWrapAlgorithm, SignatureAlgorithm
+    from ._internal import Key as _Key
 
 
-class CryptographyClient(AsyncKeyVaultClientBase):
+class CryptographyClient(KeyVaultClientBase):
     """
     Performs cryptographic operations using Azure Key Vault keys.
 
     :param key:
         Either a :class:`~azure.keyvault.keys.models.KeyVaultKey` instance as returned by
-        :func:`~azure.keyvault.keys.aio.KeyClient.get_key`, or a string.
+        :func:`~azure.keyvault.keys.KeyClient.get_key`, or a string.
         If a string, the value must be the full identifier of an Azure Key Vault key with a version.
     :type key: str or :class:`~azure.keyvault.keys.models.KeyVaultKey`
     :param credential: An object which can provide an access token for the vault, such as a credential from
-        :mod:`azure.identity.aio`
+        :mod:`azure.identity`
 
     Keyword arguments
         - **api_version** - version of the Key Vault API to use. Defaults to the most recent.
@@ -42,25 +43,25 @@ class CryptographyClient(AsyncKeyVaultClientBase):
 
     .. code-block:: python
 
-        from azure.identity.aio import DefaultAzureCredential
-        from azure.keyvault.keys.crypto.aio import CryptographyClient
+        from azure.identity import DefaultAzureCredential
+        from azure.keyvault.keys.crypto import CryptographyClient
 
         credential = DefaultAzureCredential()
 
         # create a CryptographyClient using a Key instance
-        key = await key_client.get_key("mykey")
+        key = key_client.get_key("mykey")
         crypto_client = CryptographyClient(key, credential)
 
         # or a Key's id, which must include a version
         key_id = "https://<your vault>.vault.azure.net/keys/mykey/fe4fdcab688c479a9aa80f01ffeac26"
         crypto_client = CryptographyClient(key_id, credential)
 
-    You can also obtain a ``CryptographyClient`` from a :class:`~azure.keyvault.keys.aio.KeyClient`:
+    You can also obtain a ``CryptographyClient`` from a :class:`~azure.keyvault.keys.KeyClient`:
 
     .. code-block:: python
 
-        from azure.identity.aio import DefaultAzureCredential
-        from azure.keyvault.keys.aio import KeyClient
+        from azure.identity import DefaultAzureCredential
+        from azure.keyvault.keys import KeyClient
 
         credential = DefaultAzureCredential()
         key_client = KeyClient(vault_endpoint=<your vault url>, credential=credential)
@@ -68,12 +69,14 @@ class CryptographyClient(AsyncKeyVaultClientBase):
 
     """
 
-    def __init__(self, key: "Union[KeyVaultKey, str]", credential: "TokenCredential", **kwargs: "**Any") -> None:
+    def __init__(self, key, credential, **kwargs):
+        # type: (Union[KeyVaultKey, str], TokenCredential, **Any) -> None
+
         if isinstance(key, KeyVaultKey):
             self._key = key
             self._key_id = parse_vault_id(key.id)
             self._allowed_ops = frozenset(self._key.key_operations)
-        elif isinstance(key, str):
+        elif isinstance(key, six.text_type):
             self._key = None
             self._key_id = parse_vault_id(key)
             self._keys_get_forbidden = None  # type: Optional[bool]
@@ -93,7 +96,8 @@ class CryptographyClient(AsyncKeyVaultClientBase):
         )
 
     @property
-    def key_id(self) -> str:
+    def key_id(self):
+        # type: () -> str
         """
         The full identifier of the client's key.
 
@@ -101,18 +105,19 @@ class CryptographyClient(AsyncKeyVaultClientBase):
         """
         return "/".join(self._key_id)
 
-    @distributed_trace_async
-    async def _get_key(self, **kwargs: "Any") -> "Optional[KeyVaultKey]":
+    @distributed_trace
+    def _get_key(self, **kwargs):
+        # type: (**Any) -> Optional[KeyVaultKey]
         """
         Get the client's :class:`~azure.keyvault.keys.models.KeyVaultKey`.
-        Can be `None`, if the client lacks keys/get permission.
+        Can be ``None``, if the client lacks keys/get permission.
 
-        :rtype: :class:`~azure.keyvault.keys.models.KeyVaultKey` or None
+        :rtype: :class:`~azure.keyvault.keys.models.KeyVaultKey` or ``None``
         """
 
         if not (self._key or self._keys_get_forbidden):
             try:
-                self._key = await self._client.get_key(
+                self._key = self._client.get_key(
                     self._key_id.vault_endpoint, self._key_id.name, self._key_id.version, **kwargs
                 )
                 self._allowed_ops = frozenset(self._key.key_operations)
@@ -122,12 +127,13 @@ class CryptographyClient(AsyncKeyVaultClientBase):
                 self._keys_get_forbidden = ex.status_code == 403
         return self._key
 
-    async def _get_local_key(self, **kwargs: "Any") -> "Optional[_Key]":
+    def _get_local_key(self, **kwargs):
+        # type: () -> Optional[_Key]
         """Gets an object implementing local operations. Will be ``None``, if the client was instantiated with a key
         id and lacks keys/get permission."""
 
         if not self._internal_key:
-            key = await self._get_key(**kwargs)
+            key = self._get_key(**kwargs)
             if not key:
                 return None
 
@@ -138,8 +144,9 @@ class CryptographyClient(AsyncKeyVaultClientBase):
 
         return self._internal_key
 
-    @distributed_trace_async
-    async def encrypt(self, algorithm: "EncryptionAlgorithm", plaintext: bytes, **kwargs: "Any") -> EncryptResult:
+    @distributed_trace
+    def encrypt(self, algorithm, plaintext, **kwargs):
+        # type: (EncryptionAlgorithm, bytes, **Any) -> EncryptResult
         # pylint:disable=line-too-long
         """
         Encrypt bytes using the client's key. Requires the keys/encrypt permission.
@@ -165,19 +172,25 @@ class CryptographyClient(AsyncKeyVaultClientBase):
 
         """
 
-        local_key = await self._get_local_key(**kwargs)
+        local_key = self._get_local_key(**kwargs)
         if local_key:
             if "encrypt" not in self._allowed_ops:
                 raise AzureError("This client doesn't have 'keys/encrypt' permission")
             result = local_key.encrypt(plaintext, algorithm=algorithm.value)
         else:
-            result = await self._client.encrypt(
-                self._key_id.vault_endpoint, self._key_id.name, self._key_id.version, algorithm, plaintext, **kwargs
+            result = self._client.encrypt(
+                vault_base_url=self._key_id.vault_endpoint,
+                key_name=self._key_id.name,
+                key_version=self._key_id.version,
+                algorithm=algorithm,
+                value=plaintext,
+                **kwargs
             ).result
         return EncryptResult(key_id=self.key_id, algorithm=algorithm, ciphertext=result)
 
-    @distributed_trace_async
-    async def decrypt(self, algorithm: "EncryptionAlgorithm", ciphertext: bytes, **kwargs: "**Any") -> DecryptResult:
+    @distributed_trace
+    def decrypt(self, algorithm, ciphertext, **kwargs):
+        # type: (EncryptionAlgorithm, bytes, **Any) -> DecryptResult
         """
         Decrypt a single block of encrypted data using the client's key. Requires the keys/decrypt permission.
 
@@ -194,11 +207,11 @@ class CryptographyClient(AsyncKeyVaultClientBase):
 
             from azure.keyvault.keys.crypto import EncryptionAlgorithm
 
-            result = await client.decrypt(EncryptionAlgorithm.rsa_oaep, ciphertext)
+            result = client.decrypt(EncryptionAlgorithm.rsa_oaep, ciphertext)
             print(result.decrypted_bytes)
 
         """
-        result = await self._client.decrypt(
+        result = self._client.decrypt(
             vault_base_url=self._key_id.vault_endpoint,
             key_name=self._key_id.name,
             key_version=self._key_id.version,
@@ -208,8 +221,9 @@ class CryptographyClient(AsyncKeyVaultClientBase):
         )
         return DecryptResult(decrypted_bytes=result.result)
 
-    @distributed_trace_async
-    async def wrap_key(self, algorithm: "KeyWrapAlgorithm", key: bytes, **kwargs: "Any") -> WrapKeyResult:
+    @distributed_trace
+    def wrap_key(self, algorithm, key, **kwargs):
+        # type: (KeyWrapAlgorithm, bytes, **Any) -> WrapKeyResult
         """
         Wrap a key with the client's key. Requires the keys/wrapKey permission.
 
@@ -225,20 +239,20 @@ class CryptographyClient(AsyncKeyVaultClientBase):
             from azure.keyvault.keys.crypto import KeyWrapAlgorithm
 
             # the result holds the encrypted key and identifies the encryption key and algorithm used
-            result = await client.wrap_key(KeyWrapAlgorithm.rsa_oaep, key_bytes)
+            result = client.wrap_key(KeyWrapAlgorithm.rsa_oaep, key_bytes)
             encrypted_key = result.encrypted_key
             print(result.key_id)
             print(result.algorithm)
 
         """
 
-        local_key = await self._get_local_key(**kwargs)
+        local_key = self._get_local_key(**kwargs)
         if local_key:
             if "wrapKey" not in self._allowed_ops:
                 raise AzureError("This client doesn't have 'keys/wrapKey' permission")
             result = local_key.wrap_key(key, algorithm=algorithm.value)
         else:
-            result = await self._client.wrap_key(
+            result = self._client.wrap_key(
                 self._key_id.vault_endpoint,
                 self._key_id.name,
                 self._key_id.version,
@@ -246,10 +260,12 @@ class CryptographyClient(AsyncKeyVaultClientBase):
                 value=key,
                 **kwargs
             ).result
+
         return WrapKeyResult(key_id=self.key_id, algorithm=algorithm, encrypted_key=result)
 
-    @distributed_trace_async
-    async def unwrap_key(self, algorithm: "KeyWrapAlgorithm", encrypted_key: bytes, **kwargs: "Any") -> UnwrapKeyResult:
+    @distributed_trace
+    def unwrap_key(self, algorithm, encrypted_key, **kwargs):
+        # type: (KeyWrapAlgorithm, bytes, **Any) -> UnwrapKeyResult
         """
         Unwrap a key previously wrapped with the client's key. Requires the keys/unwrapKey permission.
 
@@ -264,23 +280,24 @@ class CryptographyClient(AsyncKeyVaultClientBase):
 
             from azure.keyvault.keys.crypto import KeyWrapAlgorithm
 
-            result = await client.unwrap_key(KeyWrapAlgorithm.rsa_oaep, wrapped_bytes)
+            result = client.unwrap_key(KeyWrapAlgorithm.rsa_oaep, wrapped_bytes)
             unwrapped_bytes = result.unwrapped_bytes
 
         """
 
-        result = await self._client.unwrap_key(
-            self._key_id.vault_endpoint,
-            self._key_id.name,
-            self._key_id.version,
+        result = self._client.unwrap_key(
+            vault_base_url=self._key_id.vault_endpoint,
+            key_name=self._key_id.name,
+            key_version=self._key_id.version,
             algorithm=algorithm,
             value=encrypted_key,
             **kwargs
         )
         return UnwrapKeyResult(unwrapped_bytes=result.result)
 
-    @distributed_trace_async
-    async def sign(self, algorithm: "SignatureAlgorithm", digest: bytes, **kwargs: "**Any") -> SignResult:
+    @distributed_trace
+    def sign(self, algorithm, digest, **kwargs):
+        # type: (SignatureAlgorithm, bytes, **Any) -> SignResult
         """
         Create a signature from a digest using the client's key. Requires the keys/sign permission.
 
@@ -299,7 +316,7 @@ class CryptographyClient(AsyncKeyVaultClientBase):
             digest = hashlib.sha256(b"plaintext").digest()
 
             # sign returns a tuple with the signature and the metadata required to verify it
-            result = await client.sign(SignatureAlgorithm.rs256, digest)
+            result = client.sign(SignatureAlgorithm.rs256, digest)
 
             # the result contains the signature and identifies the key and algorithm used
             signature = result.signature
@@ -308,7 +325,7 @@ class CryptographyClient(AsyncKeyVaultClientBase):
 
         """
 
-        result = await self._client.sign(
+        result = self._client.sign(
             vault_base_url=self._key_id.vault_endpoint,
             key_name=self._key_id.name,
             key_version=self._key_id.version,
@@ -318,10 +335,9 @@ class CryptographyClient(AsyncKeyVaultClientBase):
         )
         return SignResult(key_id=self.key_id, algorithm=algorithm, signature=result.result)
 
-    @distributed_trace_async
-    async def verify(
-        self, algorithm: "SignatureAlgorithm", digest: bytes, signature: bytes, **kwargs: "**Any"
-    ) -> VerifyResult:
+    @distributed_trace
+    def verify(self, algorithm, digest, signature, **kwargs):
+        # type: (SignatureAlgorithm, bytes, bytes, **Any) -> VerifyResult
         """
         Verify a signature using the client's key. Requires the keys/verify permission.
 
@@ -337,18 +353,17 @@ class CryptographyClient(AsyncKeyVaultClientBase):
 
             from azure.keyvault.keys.crypto import SignatureAlgorithm
 
-            verified = await client.verify(SignatureAlgorithm.rs256, digest, signature)
+            verified = client.verify(SignatureAlgorithm.rs256, digest, signature)
             assert verified.result is True
 
         """
-
-        local_key = await self._get_local_key(**kwargs)
+        local_key = self._get_local_key(**kwargs)
         if local_key:
             if "verify" not in self._allowed_ops:
                 raise AzureError("This client doesn't have 'keys/verify' permission")
             result = local_key.verify(digest, signature, algorithm=algorithm.value)
         else:
-            result = await self._client.verify(
+            result = self._client.verify(
                 vault_base_url=self._key_id.vault_endpoint,
                 key_name=self._key_id.name,
                 key_version=self._key_id.version,
