+++ sdk/keyvault/azure-keyvault-keys/tests/test_keys_async.py
@@ -11,7 +11,7 @@ from devtools_testutils import ResourceGroupPreparer
 from keys_async_preparer import AsyncVaultClientPreparer
 from keys_async_test_case import AsyncKeyVaultTestCase
 
-from azure.keyvault.keys.models import JsonWebKey
+from azure.keyvault.keys import JsonWebKey
 
 from dateutil import parser as date_parse
