+++ sdk/keyvault/azure-keyvault-secrets/azure/keyvault/secrets/__init__.py
@@ -2,6 +2,6 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 # ------------------------------------
-from .client import SecretClient
+from ._client import SecretClient
 
 __all__ = ["SecretClient"]
